refactor: Push Notifications
Setup a dedicated push notification dispatcher that can be used to try a number of dynamic push notification services to send a message. Moved the Pushify API to an external section in the core package so that it can be setup as a handler to be tested when sending responses. Also added a more basic GMail push notification handler to send the final information as an email